requirements.txt
django
djangorestframework
coverage
pyjwt